normalize by plate

+++ main.py
@@ -7,7 +7,7 @@ import torch.nn as nn
 import torch.optim as optim
 import torch.utils.data as D
 
-from data_loader import data_label_split, dmso_taxol_ProfileBag, generate_data_set
+from data_loader import data_label_split, dmso_taxol_ProfileBag, generate_data_set, normalize_by_group
 from model import FullDeepSet, SmallDeepSet, profile_AttSet
 from set_experiment import mini_noise_signal_cv, test, train
 
@@ -98,18 +98,22 @@ def initialize_model(model, opt):
     return opt
 
 
-os.chdir("/bcm-share-2_3/michigan/att_pooling")
-
-file_name = "moa_data_drop_NA.csv"  # "week1_full_data.csv"
-drop_NA_data = pd.read_csv(file_name, index_col=0)
-print("Data loaded")
-data = drop_NA_data
+data_path = 'moa_data_drop_NA.csv'
+drop_NA_data=pd.read_csv(data_path, index_col=0)
+X, y = data_label_split(drop_NA_data)
+X['Metadata_PlateID_Nuclei'] = drop_NA_data['Metadata_PlateID_Nuclei'].tolist()
+X = normalize_by_group(X,'Metadata_PlateID_Nuclei')
+X.dropna('columns',inplace=True)
+X['compound'] = drop_NA_data['compound'].tolist()
+data = X
 
 feature_size = len(data_label_split(data)[0].columns)
 for i in range(args.start, args.end, 5):
     # define model
-    model = FullDeepSet(feature_size, args.pool, args.thres)
-    #     model = profile_AttSet(481,args.thres)
+    if args.pool == 'att':
+        model = profile_AttSet(feature_size, args.thres)
+    else:
+        model = FullDeepSet(feature_size, args.pool, args.thres)
     if args.cuda:
         model.cuda()
 
@@ -121,19 +125,11 @@ for i in range(args.start, args.end, 5):
         i,
         i + 1,
         data,
-        args.num_bags_train,
-        args.mean_bag_length,
-        args.var_bag_length,
         "taxol",
         "DMSO",
-        args.batch_size,
         model,
-        args.lr,
-        args.reg,
-        args.splits,
-        args.epochs,
+        args
     )
-    # feature_size = len(data_label_split(data)[0].columns)
 
     results = pd.DataFrame.from_dict(results, orient="index")
 
@@ -150,14 +146,7 @@ for i in range(args.start, args.end, 5):
         "std_pred_score_treatment",
     ]
 
-    #     if os.path.exists("deepset_att%.1f_bags%d*%d_bagsize%d_feature%d.csv"%(args.thres, args.num_bags_train, args.batch_size, args.mean_bag_length, feature_size)):
-    #         results.to_csv("deepset_att%.1f_bags%d*%d_bagsize%d_feature%d.csv"%(args.thres, args.num_bags_train, args.batch_size, args.mean_bag_length, feature_size), mode='a', header=False)
-    #     else:
-    #         results.to_csv("deepset_att%.1f_bags%d*%d_bagsize%d_feature%d.csv"%(args.thres, args.num_bags_train, args.batch_size, args.mean_bag_length, feature_size))
-
-    if os.path.exists(
-        "deepset_%s%.1f_bags%d*%d_bagsize%d_feature%d.csv"
-        % (
+    res_path = "deepset_%s%.1f_bags%d*%d_bagsize%d_feature%d.csv"% (
             args.pool,
             args.thres,
             args.num_bags_train,
@@ -165,29 +154,12 @@ for i in range(args.start, args.end, 5):
             args.mean_bag_length,
             feature_size,
         )
-    ):
+    
+    if os.path.exists(res_path):
         results.to_csv(
-            "deepset_%s%.1f_bags%d*%d_bagsize%d_feature%d.csv"
-            % (
-                args.pool,
-                args.thres,
-                args.num_bags_train,
-                args.batch_size,
-                args.mean_bag_length,
-                feature_size,
-            ),
+            res_path,
             mode="a",
             header=False,
         )
     else:
-        results.to_csv(
-            "deepset_%s%.1f_bags%d*%d_bagsize%d_feature%d.csv"
-            % (
-                args.pool,
-                args.thres,
-                args.num_bags_train,
-                args.batch_size,
-                args.mean_bag_length,
-                feature_size,
-            )
-        )
+        results.to_csv(res_path)
